move value_counts, pct_change, explode, agg
