+++ python/cudf/cudf/tests/dataframe/methods/test_rename.py
@@ -76,3 +76,74 @@ def test_rename_reset_label_dtype():
     result = cudf.DataFrame(data).rename(columns=col_mapping)
     expected = pd.DataFrame(data).rename(columns=col_mapping)
     assert_eq(result, expected)
+
+
+def test_dataframe_rename_columns_keep_type():
+    gdf = cudf.DataFrame([[1, 2, 3]])
+    gdf.columns = cudf.Index([4, 5, 6], dtype=np.int8)
+    result = gdf.rename({4: 50}, axis="columns").columns
+    expected = pd.Index([50, 5, 6], dtype=np.int8)
+    assert_eq(result, expected)
+
+
+def test_dataframe_rename_duplicate_column():
+    gdf = cudf.DataFrame({"a": [1, 2, 3], "b": [3, 4, 5]})
+    with pytest.raises(
+        ValueError, match="Duplicate column names are not allowed"
+    ):
+        gdf.rename(columns={"a": "b"}, inplace=True)
+
+
+@pytest.mark.parametrize("level", ["x", 0])
+def test_rename_for_level_MultiIndex_dataframe(level):
+    data = {"a": [1, 2, 3], "b": [4, 5, 6], "c": [7, 8, 9]}
+    index = {0: 123, 1: 4, 2: 6}
+    pdf = pd.DataFrame(
+        data,
+        index=pd.MultiIndex.from_tuples([(0, 1, 2), (1, 2, 3), (2, 3, 4)]),
+    )
+    pdf.index.names = ["x", "y", "z"]
+    gdf = cudf.from_pandas(pdf)
+
+    expect = pdf.rename(index=index, level=level)
+    got = gdf.rename(index=index, level=level)
+
+    assert_eq(expect, got)
+
+
+@pytest.mark.parametrize(
+    "columns",
+    [{"a": "f", "b": "g"}, {1: 3, 2: 4}, lambda s: 2 * s],
+)
+@pytest.mark.parametrize("level", [0, 1])
+def test_rename_for_level_MultiColumn_dataframe(columns, level):
+    gdf = cudf.DataFrame({"a": [1, 2, 3], "b": [4, 5, 6], "c": [7, 8, 9]})
+    gdf.columns = pd.MultiIndex.from_tuples([("a", 1), ("a", 2), ("b", 1)])
+
+    pdf = gdf.to_pandas()
+
+    expect = pdf.rename(columns=columns, level=level)
+    got = gdf.rename(columns=columns, level=level)
+
+    assert_eq(expect, got)
+
+
+def test_rename_for_level_RangeIndex_dataframe():
+    gdf = cudf.DataFrame({"a": [1, 2, 3], "b": [4, 5, 6], "c": [7, 8, 9]})
+    pdf = gdf.to_pandas()
+
+    expect = pdf.rename(columns={"a": "f"}, index={0: 3, 1: 4}, level=0)
+    got = gdf.rename(columns={"a": "f"}, index={0: 3, 1: 4}, level=0)
+
+    assert_eq(expect, got)
+
+
+def test_rename_for_level_is_None_MC():
+    gdf = cudf.DataFrame({"a": [1, 2, 3], "b": [4, 5, 6], "c": [7, 8, 9]})
+    gdf.columns = pd.MultiIndex.from_tuples([("a", 1), ("a", 2), ("b", 1)])
+    pdf = gdf.to_pandas()
+
+    expect = pdf.rename(columns={"a": "f"}, level=None)
+    got = gdf.rename(columns={"a": "f"}, level=None)
+
+    assert_eq(expect, got)
