+++ python/cudf/cudf/tests/indexes/rangeindex/methods/test_nunique.py
@@ -1,0 +1,15 @@
+# Copyright (c) 2025, NVIDIA CORPORATION.
+
+import pandas as pd
+
+import cudf
+
+
+def test_nunique():
+    gidx = cudf.RangeIndex(5)
+    pidx = pd.RangeIndex(5)
+
+    actual = gidx.nunique()
+    expected = pidx.nunique()
+
+    assert actual == expected
