+++ python/cudf/cudf/tests/test_dataframe.py
@@ -17,7 +17,6 @@ import pytest
 from packaging import version
 
 import cudf
-from cudf.api.extensions import no_default
 from cudf.core._compat import (
     PANDAS_CURRENT_SUPPORTED_VERSION,
     PANDAS_VERSION,
@@ -4682,819 +4681,3 @@ def test_dataframe_constructor_column_index_only():
     assert not id(gdf["a"]._column) == id(gdf["b"]._column) and not id(
         gdf["b"]._column
     ) == id(gdf["c"]._column)
-
-
-@pytest_unmark_spilling
-@pytest.mark.parametrize(
-    "data",
-    [
-        {"a": [1, 2.5, 3], "b": [3, 4.5, 5], "c": [2.0, 3.0, 4.0]},
-        {"a": [1, 2.2, 3], "b": [2.0, 3.0, 4.0], "c": [5.0, 6.0, 4.0]},
-    ],
-)
-@pytest.mark.parametrize(
-    "aggs",
-    [
-        ["min", "sum", "max"],
-        ("min", "sum", "max"),
-        {"min", "sum", "max"},
-        "sum",
-        {"a": "sum", "b": "min", "c": "max"},
-        {"a": ["sum"], "b": ["min"], "c": ["max"]},
-        {"a": ("sum"), "b": ("min"), "c": ("max")},
-        {"a": {"sum"}, "b": {"min"}, "c": {"max"}},
-        {"a": ["sum", "min"], "b": ["sum", "max"], "c": ["min", "max"]},
-        {"a": ("sum", "min"), "b": ("sum", "max"), "c": ("min", "max")},
-        {"a": {"sum", "min"}, "b": {"sum", "max"}, "c": {"min", "max"}},
-    ],
-)
-def test_agg_for_dataframes(data, aggs):
-    pdf = pd.DataFrame(data)
-    gdf = cudf.DataFrame(data)
-
-    expect = pdf.agg(aggs).sort_index()
-    got = gdf.agg(aggs).sort_index()
-
-    assert_eq(expect, got, check_dtype=True)
-
-
-@pytest_unmark_spilling
-@pytest.mark.parametrize(
-    "data",
-    [
-        {"a": [1, 2, 3], "b": [3.0, 4.0, 5.0], "c": [True, True, False]},
-        {"a": [1, 2, 3], "b": [True, True, False], "c": [False, True, False]},
-    ],
-)
-@pytest.mark.parametrize(
-    "aggs",
-    [
-        ["min", "sum", "max"],
-        "sum",
-        {"a": "sum", "b": "min", "c": "max"},
-    ],
-)
-def test_agg_for_dataframes_error(data, aggs):
-    gdf = cudf.DataFrame(data)
-
-    with pytest.raises(TypeError):
-        gdf.agg(aggs)
-
-
-def test_agg_for_unsupported_function():
-    gdf = cudf.DataFrame({"a": [1, 2, 3], "b": [3.0, 4.0, 5.0]})
-
-    with pytest.raises(NotImplementedError):
-        gdf.agg({"a": np.sum, "b": np.min, "c": np.max})
-
-
-def test_agg_for_dataframe_with_invalid_function():
-    aggs = "asdf"
-    gdf = cudf.DataFrame({"a": [1, 2, 3], "b": [3.0, 4.0, 5.0]})
-
-    with pytest.raises(
-        AttributeError,
-        match=f"{aggs} is not a valid function for 'DataFrame' object",
-    ):
-        gdf.agg(aggs)
-
-
-def test_agg_for_series_with_invalid_function():
-    aggs = {"a": "asdf"}
-    gdf = cudf.DataFrame({"a": [1, 2, 3], "b": [3.0, 4.0, 5.0]})
-
-    with pytest.raises(
-        AttributeError,
-        match=f"{aggs['a']} is not a valid function for 'Series' object",
-    ):
-        gdf.agg(aggs)
-
-
-@pytest.mark.parametrize(
-    "aggs",
-    [
-        "sum",
-        ["min", "sum", "max"],
-        {"a": {"sum", "min"}, "b": {"sum", "max"}, "c": {"min", "max"}},
-    ],
-)
-def test_agg_for_dataframe_with_string_columns(aggs):
-    gdf = cudf.DataFrame(
-        {"a": ["m", "n", "o"], "b": ["t", "u", "v"], "c": ["x", "y", "z"]},
-        index=["a", "b", "c"],
-    )
-
-    with pytest.raises(
-        NotImplementedError,
-        match=re.escape(
-            "DataFrame.agg() is not supported for "
-            "frames containing string columns"
-        ),
-    ):
-        gdf.agg(aggs)
-
-
-@pytest_unmark_spilling
-@pytest.mark.parametrize("overwrite", [True, False])
-@pytest.mark.parametrize(
-    "left_keys,right_keys",
-    [
-        [("a", "b"), ("a", "b")],
-        [("a", "b"), ("a", "c")],
-        [("a", "b"), ("d", "e")],
-    ],
-)
-@pytest.mark.parametrize(
-    "data_left,data_right",
-    [
-        [([1, 2, 3], [3, 4, 5]), ([1, 2, 3], [3, 4, 5])],
-        [
-            ([1.0, 2.0, 3.0], [3.0, 4.0, 5.0]),
-            ([1.0, 2.0, 3.0], [3.0, 4.0, 5.0]),
-        ],
-        [
-            ([True, False, True], [False, False, False]),
-            ([True, False, True], [False, False, False]),
-        ],
-        [
-            ([np.nan, np.nan, np.nan], [np.nan, np.nan, np.nan]),
-            ([np.nan, np.nan, np.nan], [np.nan, np.nan, np.nan]),
-        ],
-        [([1, 2, 3], [3, 4, 5]), ([1, 2, 4], [30, 40, 50])],
-        [
-            ([1.0, 2.0, 3.0], [3.0, 4.0, 5.0]),
-            ([1.0, 2.0, 4.0], [30.0, 40.0, 50.0]),
-        ],
-        [([1, 2, 3], [3, 4, 5]), ([10, 20, 40], [30, 40, 50])],
-        [
-            ([1.0, 2.0, 3.0], [3.0, 4.0, 5.0]),
-            ([10.0, 20.0, 40.0], [30.0, 40.0, 50.0]),
-        ],
-    ],
-)
-def test_update_for_dataframes(
-    left_keys, right_keys, data_left, data_right, overwrite
-):
-    errors = "ignore"
-    join = "left"
-    left = dict(zip(left_keys, data_left, strict=True))
-    right = dict(zip(right_keys, data_right, strict=True))
-    pdf = pd.DataFrame(left)
-    gdf = cudf.DataFrame(left, nan_as_null=False)
-
-    other_pd = pd.DataFrame(right)
-    other_gd = cudf.DataFrame(right, nan_as_null=False)
-
-    pdf.update(other=other_pd, join=join, overwrite=overwrite, errors=errors)
-    gdf.update(other=other_gd, join=join, overwrite=overwrite, errors=errors)
-
-    assert_eq(pdf, gdf, check_dtype=False)
-
-
-def test_update_for_right_join():
-    gdf = cudf.DataFrame({"a": [1, 2, 3], "b": [3.0, 4.0, 5.0]})
-    other_gd = cudf.DataFrame({"a": [1, np.nan, 3], "b": [np.nan, 2.0, 5.0]})
-
-    with pytest.raises(
-        NotImplementedError, match="Only left join is supported"
-    ):
-        gdf.update(other_gd, join="right")
-
-
-def test_update_for_data_overlap():
-    errors = "raise"
-    pdf = pd.DataFrame({"a": [1, 2, 3], "b": [3.0, 4.0, 5.0]})
-    gdf = cudf.DataFrame({"a": [1, 2, 3], "b": [3.0, 4.0, 5.0]})
-
-    other_pd = pd.DataFrame({"a": [1, np.nan, 3], "b": [np.nan, 2.0, 5.0]})
-    other_gd = cudf.DataFrame({"a": [1, np.nan, 3], "b": [np.nan, 2.0, 5.0]})
-
-    assert_exceptions_equal(
-        lfunc=pdf.update,
-        rfunc=gdf.update,
-        lfunc_args_and_kwargs=([other_pd, errors], {}),
-        rfunc_args_and_kwargs=([other_gd, errors], {}),
-    )
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        {"a": [[1], [2], [3]]},
-        {
-            "left-a": [0, 1, 2],
-            "a": [[1], None, [3]],
-            "right-a": ["abc", "def", "ghi"],
-        },
-        {
-            "left-a": [[], None, None],
-            "a": [[1], None, [3]],
-            "right-a": ["abc", "def", "ghi"],
-        },
-    ],
-)
-def test_dataframe_roundtrip_arrow_list_dtype(data):
-    gdf = cudf.DataFrame(data)
-    table = gdf.to_arrow()
-    expected = cudf.DataFrame.from_arrow(table)
-
-    assert_eq(gdf, expected)
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        {"a": [{"one": 3, "two": 4, "three": 10}]},
-        {
-            "left-a": [0, 1, 2],
-            "a": [{"x": 0.23, "y": 43}, None, {"x": 23.9, "y": 4.3}],
-            "right-a": ["abc", "def", "ghi"],
-        },
-        {
-            "left-a": [{"a": 1}, None, None],
-            "a": [
-                {"one": 324, "two": 23432, "three": 324},
-                None,
-                {"one": 3.24, "two": 1, "three": 324},
-            ],
-            "right-a": ["abc", "def", "ghi"],
-        },
-    ],
-)
-def test_dataframe_roundtrip_arrow_struct_dtype(data):
-    gdf = cudf.DataFrame(data)
-    table = gdf.to_arrow()
-    expected = cudf.DataFrame.from_arrow(table)
-
-    assert_eq(gdf, expected)
-
-
-def test_dataframe_setitem_cupy_array():
-    rng = np.random.default_rng(seed=0)
-    pdf = pd.DataFrame(rng.standard_normal(size=(10, 2)))
-    gdf = cudf.from_pandas(pdf)
-
-    gpu_array = cupy.array([True, False] * 5)
-    pdf[gpu_array.get()] = 1.5
-    gdf[gpu_array] = 1.5
-
-    assert_eq(pdf, gdf)
-
-
-@pytest.mark.parametrize("level", ["x", 0])
-def test_rename_for_level_MultiIndex_dataframe(level):
-    data = {"a": [1, 2, 3], "b": [4, 5, 6], "c": [7, 8, 9]}
-    index = {0: 123, 1: 4, 2: 6}
-    pdf = pd.DataFrame(
-        data,
-        index=pd.MultiIndex.from_tuples([(0, 1, 2), (1, 2, 3), (2, 3, 4)]),
-    )
-    pdf.index.names = ["x", "y", "z"]
-    gdf = cudf.from_pandas(pdf)
-
-    expect = pdf.rename(index=index, level=level)
-    got = gdf.rename(index=index, level=level)
-
-    assert_eq(expect, got)
-
-
-@pytest.mark.parametrize(
-    "columns",
-    [{"a": "f", "b": "g"}, {1: 3, 2: 4}, lambda s: 2 * s],
-)
-@pytest.mark.parametrize("level", [0, 1])
-def test_rename_for_level_MultiColumn_dataframe(columns, level):
-    gdf = cudf.DataFrame({"a": [1, 2, 3], "b": [4, 5, 6], "c": [7, 8, 9]})
-    gdf.columns = pd.MultiIndex.from_tuples([("a", 1), ("a", 2), ("b", 1)])
-
-    pdf = gdf.to_pandas()
-
-    expect = pdf.rename(columns=columns, level=level)
-    got = gdf.rename(columns=columns, level=level)
-
-    assert_eq(expect, got)
-
-
-def test_rename_for_level_RangeIndex_dataframe():
-    gdf = cudf.DataFrame({"a": [1, 2, 3], "b": [4, 5, 6], "c": [7, 8, 9]})
-    pdf = gdf.to_pandas()
-
-    expect = pdf.rename(columns={"a": "f"}, index={0: 3, 1: 4}, level=0)
-    got = gdf.rename(columns={"a": "f"}, index={0: 3, 1: 4}, level=0)
-
-    assert_eq(expect, got)
-
-
-def test_rename_for_level_is_None_MC():
-    gdf = cudf.DataFrame({"a": [1, 2, 3], "b": [4, 5, 6], "c": [7, 8, 9]})
-    gdf.columns = pd.MultiIndex.from_tuples([("a", 1), ("a", 2), ("b", 1)])
-    pdf = gdf.to_pandas()
-
-    expect = pdf.rename(columns={"a": "f"}, level=None)
-    got = gdf.rename(columns={"a": "f"}, level=None)
-
-    assert_eq(expect, got)
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        [
-            [[1, 2, 3], 11, "a"],
-            [None, 22, "e"],
-            [[4], 33, "i"],
-            [[], 44, "o"],
-            [[5, 6], 55, "u"],
-        ],  # nested
-        [
-            [1, 11, "a"],
-            [2, 22, "e"],
-            [3, 33, "i"],
-            [4, 44, "o"],
-            [5, 55, "u"],
-        ],  # non-nested
-    ],
-)
-@pytest.mark.parametrize(
-    ("labels", "label_to_explode"),
-    [
-        (None, 0),
-        (pd.Index(["a", "b", "c"]), "a"),
-        (
-            pd.MultiIndex.from_tuples(
-                [(0, "a"), (0, "b"), (1, "a")], names=["l0", "l1"]
-            ),
-            (0, "a"),
-        ),
-    ],
-)
-@pytest.mark.parametrize("ignore_index", [True, False])
-@pytest.mark.parametrize(
-    "p_index",
-    [
-        None,
-        ["ia", "ib", "ic", "id", "ie"],
-        pd.MultiIndex.from_tuples(
-            [(0, "a"), (0, "b"), (0, "c"), (1, "a"), (1, "b")]
-        ),
-    ],
-)
-def test_explode(data, labels, ignore_index, p_index, label_to_explode):
-    pdf = pd.DataFrame(data, index=p_index, columns=labels)
-    gdf = cudf.from_pandas(pdf)
-
-    expect = pdf.explode(label_to_explode, ignore_index)
-    got = gdf.explode(label_to_explode, ignore_index)
-
-    assert_eq(expect, got, check_dtype=False)
-
-
-def test_explode_preserve_categorical():
-    gdf = cudf.DataFrame(
-        {
-            "A": [[1, 2], None, [2, 3]],
-            "B": cudf.Series([0, 1, 2], dtype="category"),
-        }
-    )
-    result = gdf.explode("A")
-    expected = cudf.DataFrame(
-        {
-            "A": [1, 2, None, 2, 3],
-            "B": cudf.Series([0, 0, 1, 2, 2], dtype="category"),
-        }
-    )
-    expected.index = cudf.Index([0, 0, 1, 2, 2])
-    assert_eq(result, expected)
-
-
-@pytest.mark.parametrize(
-    "data,ascending,expected_data",
-    [
-        (
-            {"a": [10, 0, 2], "b": [-10, 10, 1]},
-            True,
-            [1, 2, 0],
-        ),
-        (
-            {"a": [10, 0, 2], "b": [-10, 10, 1]},
-            False,
-            [0, 2, 1],
-        ),
-    ],
-)
-def test_dataframe_argsort(data, ascending, expected_data):
-    actual = cudf.DataFrame(data).argsort(ascending=ascending)
-    expected = cupy.array(expected_data, dtype="int32")
-
-    assert_eq(actual, expected)
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        np.random.RandomState(seed=10).randint(-50, 50, (25, 30)),
-        np.random.RandomState(seed=10).random_sample((4, 4)),
-        np.array([1.123, 2.343, 5.890, 0.0]),
-        [True, False, True, False, False],
-        {"a": [1.123, 2.343, np.nan, np.nan], "b": [None, 3, 9.08, None]},
-    ],
-)
-@pytest.mark.parametrize("periods", (-5, -1, 0, 1, 5))
-def test_diff_numeric_dtypes(data, periods):
-    gdf = cudf.DataFrame(data)
-    pdf = gdf.to_pandas()
-
-    actual = gdf.diff(periods=periods, axis=0)
-    expected = pdf.diff(periods=periods, axis=0)
-
-    assert_eq(
-        expected,
-        actual,
-        check_dtype=False,
-    )
-
-
-@pytest.mark.parametrize(
-    ("precision", "scale"),
-    [(5, 2), (8, 5)],
-)
-@pytest.mark.parametrize(
-    "dtype",
-    [cudf.Decimal32Dtype, cudf.Decimal64Dtype],
-)
-def test_diff_decimal_dtypes(precision, scale, dtype):
-    gdf = cudf.DataFrame(
-        np.random.default_rng(seed=42).uniform(10.5, 75.5, (10, 6)),
-        dtype=dtype(precision=precision, scale=scale),
-    )
-    pdf = gdf.to_pandas()
-
-    actual = gdf.diff()
-    expected = pdf.diff()
-
-    assert_eq(
-        expected,
-        actual,
-        check_dtype=False,
-    )
-
-
-def test_diff_invalid_axis():
-    gdf = cudf.DataFrame(np.array([1.123, 2.343, 5.890, 0.0]))
-    with pytest.raises(NotImplementedError, match="Only axis=0 is supported."):
-        gdf.diff(periods=1, axis=1)
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        {
-            "int_col": [1, 2, 3, 4, 5],
-            "float_col": [1.0, 2.0, 3.0, 4.0, 5.0],
-            "string_col": ["a", "b", "c", "d", "e"],
-        },
-        ["a", "b", "c", "d", "e"],
-    ],
-)
-def test_diff_unsupported_dtypes(data):
-    gdf = cudf.DataFrame(data)
-    with pytest.raises(
-        TypeError,
-        match=r"unsupported operand type\(s\)",
-    ):
-        gdf.diff()
-
-
-def test_diff_many_dtypes():
-    pdf = pd.DataFrame(
-        {
-            "dates": pd.date_range("2020-01-01", "2020-01-06", freq="D"),
-            "bools": [True, True, True, False, True, True],
-            "floats": [1.0, 2.0, 3.5, np.nan, 5.0, -1.7],
-            "ints": [1, 2, 3, 3, 4, 5],
-            "nans_nulls": [np.nan, None, None, np.nan, np.nan, None],
-        }
-    )
-    gdf = cudf.from_pandas(pdf)
-    assert_eq(pdf.diff(), gdf.diff())
-    assert_eq(pdf.diff(periods=2), gdf.diff(periods=2))
-
-
-def test_dataframe_assign_cp_np_array():
-    m, n = 5, 3
-    cp_ndarray = cupy.random.randn(m, n)
-    pdf = pd.DataFrame({f"f_{i}": range(m) for i in range(n)})
-    gdf = cudf.DataFrame({f"f_{i}": range(m) for i in range(n)})
-    pdf[[f"f_{i}" for i in range(n)]] = cupy.asnumpy(cp_ndarray)
-    gdf[[f"f_{i}" for i in range(n)]] = cp_ndarray
-
-    assert_eq(pdf, gdf)
-
-
-def test_dataframe_nunique():
-    gdf = cudf.DataFrame({"a": [1, 2, 3], "b": [1, 1, 0]})
-    pdf = gdf.to_pandas()
-
-    actual = gdf.nunique()
-    expected = pdf.nunique()
-
-    assert_eq(expected, actual)
-
-
-@pytest.mark.parametrize(
-    "columns",
-    [
-        pd.RangeIndex(2, name="foo"),
-        pd.MultiIndex.from_arrays([[1, 2], [2, 3]], names=["foo", 1]),
-        pd.Index([3, 5], dtype=np.int8, name="foo"),
-    ],
-)
-def test_nunique_preserve_column_in_index(columns):
-    df = cudf.DataFrame([[1, 2]], columns=columns)
-    result = df.nunique().index.to_pandas()
-    assert_eq(result, columns, exact=True)
-
-
-def test_dataframe_nunique_index():
-    gdf = cudf.DataFrame(
-        {"key": [0, 1, 1, 0, 0, 1], "val": [1, 8, 3, 9, -3, 8]}
-    )
-    pdf = gdf.to_pandas()
-
-    actual = gdf.index.nunique()
-    expected = pdf.index.nunique()
-
-    assert_eq(expected, actual)
-
-
-def test_dataframe_rename_duplicate_column():
-    gdf = cudf.DataFrame({"a": [1, 2, 3], "b": [3, 4, 5]})
-    with pytest.raises(
-        ValueError, match="Duplicate column names are not allowed"
-    ):
-        gdf.rename(columns={"a": "b"}, inplace=True)
-
-
-def test_dataframe_rename_columns_keep_type():
-    gdf = cudf.DataFrame([[1, 2, 3]])
-    gdf.columns = cudf.Index([4, 5, 6], dtype=np.int8)
-    result = gdf.rename({4: 50}, axis="columns").columns
-    expected = pd.Index([50, 5, 6], dtype=np.int8)
-    assert_eq(result, expected)
-
-
-@pytest_unmark_spilling
-@pytest.mark.skipif(
-    PANDAS_VERSION < PANDAS_CURRENT_SUPPORTED_VERSION,
-    reason="warning not present in older pandas versions",
-)
-@pytest.mark.parametrize(
-    "data",
-    [
-        np.random.RandomState(seed=10).randint(-50, 50, (10, 10)),
-        np.random.RandomState(seed=10).random_sample((4, 4)),
-        np.array([1.123, 2.343, 5.890, 0.0]),
-        {"a": [1.123, 2.343, np.nan, np.nan], "b": [None, 3, 9.08, None]},
-    ],
-)
-@pytest.mark.parametrize("periods", [-5, 0, 2])
-@pytest.mark.parametrize(
-    "fill_method", ["ffill", "bfill", "pad", "backfill", no_default]
-)
-def test_dataframe_pct_change(data, periods, fill_method):
-    gdf = cudf.DataFrame(data)
-    pdf = gdf.to_pandas()
-
-    with expect_warning_if(fill_method is not no_default):
-        actual = gdf.pct_change(periods=periods, fill_method=fill_method)
-    with expect_warning_if(
-        fill_method is not no_default or pdf.isna().any().any()
-    ):
-        expected = pdf.pct_change(periods=periods, fill_method=fill_method)
-
-    assert_eq(expected, actual)
-
-
-@pytest.mark.parametrize("numeric_only", [True, False])
-def test_mean_timeseries(numeric_only):
-    gdf = cudf.datasets.timeseries()
-    if not numeric_only:
-        gdf = gdf.select_dtypes(include="number")
-    pdf = gdf.to_pandas()
-
-    expected = pdf.mean(numeric_only=numeric_only)
-    actual = gdf.mean(numeric_only=numeric_only)
-
-    assert_eq(expected, actual)
-
-
-@pytest.mark.parametrize("numeric_only", [True, False])
-def test_std_different_dtypes(numeric_only):
-    gdf = cudf.DataFrame(
-        {
-            "a": [1, 2, 3, 4, 5],
-            "b": ["a", "b", "c", "d", "e"],
-            "c": [1.0, 2.0, 3.0, 4.0, 5.0],
-        }
-    )
-    if not numeric_only:
-        gdf = gdf.select_dtypes(include="number")
-    pdf = gdf.to_pandas()
-
-    expected = pdf.std(numeric_only=numeric_only)
-    actual = gdf.std(numeric_only=numeric_only)
-
-    assert_eq(expected, actual)
-
-
-def test_empty_numeric_only():
-    gdf = cudf.DataFrame(
-        {
-            "id": ["a", "a", "a", "b", "b", "b", "c", "c", "c"],
-            "val1": ["v", "n", "k", "l", "m", "i", "y", "r", "w"],
-            "val2": ["d", "d", "d", "e", "e", "e", "f", "f", "f"],
-        }
-    )
-    pdf = gdf.to_pandas()
-    expected = pdf.prod(numeric_only=True)
-    actual = gdf.prod(numeric_only=True)
-    assert_eq(expected, actual, check_dtype=True)
-
-
-# Note that for now expressions do not automatically handle casting, so inputs
-# need to be casted appropriately
-@pytest.mark.parametrize(
-    "expr, dtype",
-    [
-        ("a", int),
-        ("+a", int),
-        ("a + b", int),
-        ("a == b", int),
-        ("a / b", float),
-        ("a * b", int),
-        ("a > b", int),
-        ("a >= b", int),
-        ("a > b > c", int),
-        ("a > b < c", int),
-        ("a & b", int),
-        ("a & b | c", int),
-        ("sin(a)", float),
-        ("exp(sin(abs(a)))", float),
-        ("sqrt(floor(a))", float),
-        ("ceil(arctanh(a))", float),
-        ("(a + b) - (c * d)", int),
-        ("~a", int),
-        ("(a > b) and (c > d)", int),
-        ("(a > b) or (c > d)", int),
-        ("not (a > b)", int),
-        ("a + 1", int),
-        ("a + 1.0", float),
-        ("-a + 1", int),
-        ("+a + 1", int),
-        ("e = a + 1", int),
-        (
-            """
-            e = log(cos(a)) + 1.0
-            f = abs(c) - exp(d)
-            """,
-            float,
-        ),
-        ("a_b_are_equal = (a == b)", int),
-        ("a > b", str),
-        ("a < '1'", str),
-        ('a == "1"', str),
-    ],
-)
-@pytest.mark.parametrize("nrows", [0, 10])
-def test_dataframe_eval(nrows, expr, dtype):
-    arr = np.ones(nrows)
-    df_eval = cudf.DataFrame({"a": arr, "b": arr, "c": arr, "d": arr})
-    df_eval = df_eval.astype(dtype)
-    with _hide_ufunc_warnings(expr):
-        expect = df_eval.to_pandas().eval(expr)
-    got = df_eval.eval(expr)
-    # In the specific case where the evaluated expression is a unary function
-    # of a single column with no nesting, pandas will retain the name. This
-    # level of compatibility is out of scope for now.
-    assert_eq(expect, got, check_names=False)
-
-    # Test inplace
-    if re.search("[^=><]=[^=]", expr) is not None:
-        pdf_eval = df_eval.to_pandas()
-        with _hide_ufunc_warnings(expr):
-            pdf_eval.eval(expr, inplace=True)
-        df_eval.eval(expr, inplace=True)
-        assert_eq(pdf_eval, df_eval)
-
-
-@pytest.mark.parametrize(
-    "expr",
-    [
-        """
-        e = a + b
-        a == b
-        """,
-        "a_b_are_equal = (a == b) = c",
-    ],
-)
-def test_dataframe_eval_errors(expr):
-    df = cudf.DataFrame({"a": [1, 2, 3, 4, 5], "b": [1, 2, 3, 4, 5]})
-    with pytest.raises(ValueError):
-        df.eval(expr)
-
-
-def test_dataframe_eval_misc():
-    df = cudf.DataFrame({"a": [1, 2, 3, None, 5]})
-    got = df.eval("isnull(a)")
-    assert_eq(got, cudf.Series.isnull(df["a"]), check_names=False)
-
-    df.eval("c = isnull(1)", inplace=True)
-    assert_eq(df["c"], cudf.Series([False] * len(df), name="c"))
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        {"num_legs": [2, 4, 4, 6], "num_wings": [2, 0, 0, 0]},
-        {
-            "first_name": ["John", "Anne", "John", "Beth"],
-            "middle_name": ["Smith", None, None, "Louise"],
-        },
-    ],
-)
-@pytest.mark.parametrize("sort", [True, False])
-@pytest.mark.parametrize("ascending", [True, False])
-@pytest.mark.parametrize("normalize", [True, False])
-@pytest.mark.parametrize("dropna", [True, False])
-@pytest.mark.parametrize("use_subset", [True, False])
-def test_value_counts(
-    data,
-    sort,
-    ascending,
-    normalize,
-    dropna,
-    use_subset,
-):
-    subset = [next(iter(data.keys()))]
-    gdf = cudf.DataFrame(data)
-    pdf = gdf.to_pandas()
-
-    got = gdf.value_counts(
-        subset=subset if (use_subset) else None,
-        sort=sort,
-        ascending=ascending,
-        normalize=normalize,
-        dropna=dropna,
-    )
-    expected = pdf.value_counts(
-        subset=subset if (use_subset) else None,
-        sort=sort,
-        ascending=ascending,
-        normalize=normalize,
-        dropna=dropna,
-    )
-
-    if not dropna:
-        # Convert the Pandas series to a cuDF one due to difference
-        # in the handling of NaNs between the two (<NA> in cuDF and
-        # NaN in Pandas) when dropna=False.
-        assert_eq(got.sort_index(), cudf.from_pandas(expected).sort_index())
-    else:
-        assert_eq(got.sort_index(), expected.sort_index())
-
-
-def test_value_counts_no_subset():
-    gdf = cudf.DataFrame({"a": [1, 2, 3], "b": [1, 1, 0]})
-    with pytest.raises(KeyError):
-        gdf.value_counts(subset=["not_a_column_name"])
-
-
-def test_multiindex_wildcard_selection_all():
-    midx = cudf.MultiIndex.from_tuples(
-        [(c1, c2) for c1 in "abc" for c2 in "ab"]
-    )
-    df = cudf.DataFrame({f"{i}": [i] for i in range(6)})
-    df.columns = midx
-    expect = df.to_pandas().loc[:, (slice(None), "b")]
-    got = df.loc[:, (slice(None), "b")]
-    assert_eq(expect, got)
-
-
-@pytest_xfail(reason="Not yet properly supported.")
-def test_multiindex_wildcard_selection_partial():
-    midx = cudf.MultiIndex.from_tuples(
-        [(c1, c2) for c1 in "abc" for c2 in "ab"]
-    )
-    df = cudf.DataFrame({f"{i}": [i] for i in range(6)})
-    df.columns = midx
-    expect = df.to_pandas().loc[:, (slice("a", "b"), "b")]
-    got = df.loc[:, (slice("a", "b"), "b")]
-    assert_eq(expect, got)
-
-
-def test_dataframe_assign_scalar_to_empty_series():
-    expected = pd.DataFrame({"a": []})
-    actual = cudf.DataFrame({"a": []})
-    expected.a = 0
-    actual.a = 0
-    assert_eq(expected, actual)
